resources/scripts/requirements.txt
PyYAML==6.0
requests==2.30.0
urllib3==2.0.2